requirements.txt
appdirs==1.4.3
attrs==19.3.0
autopep8==1.5.1
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
django~=2.2
django-filter==2.2.0
django-leaflet==0.26.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-gis==0.15
geographiclib==1.50
geopy==1.21.0
gprof2dot==2019.11.30
idna==2.9
Jinja2==2.11.1
MarkupSafe==1.1.1
pathspec==0.8.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
django-silk==4.0.1
ua-parser~=0.10
django-environ==0.4.5
django-simple-history==2.8.0
django-pipeline==1.7.0
